gen-ai-in-prod-book/requirements.txt
deeplake==3.6.19
transformers==4.32.0
numpy==1.23.5
trl==0.7.1
peft==0.5.0
wandb==0.15.8
accelerate==0.22.0
tqdm==4.66.1
neural_compressor===2.2.1
onnx===1.14.1
pandas==2.0.3
scipy==1.11.2

# New OpenAI SDK and tokenizer
openai>=1.99.9,<2.0.0
tiktoken>=0.7,<1

# Modern LangChain stack
langchain>=0.3.10,<1
langchain-core>=0.3.74,<1
langchain-community>=0.3.10,<1
langchain-openai==0.3.30
newspaper3k==0.2.8
lxml>=5.0.0
lxml_html_clean>=0.1.0
pypdf>=4.0.0,<5
faiss-cpu

# platform-specific (skip on Windows)
deepspeed==0.10.1 ; sys_platform != "win32"
bitsandbytes==0.41.1 ; sys_platform != "win32"

